backend/requirements.txt
Django>=5.1.5
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-jazzmin>=2.6.0
openai>=1.12.0
python-dotenv>=1.0.0 